server/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-filter
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
openai
python-decouple
requests-oauthlib
celery
schedule
django-celery-results